refactor: move config from iroh-ctl into iroh-one named iroh_ctl for the time being
