+++ iroh-ctl/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod config;
 #[cfg(feature = "testing")]
 mod fixture;
 pub mod gateway;
